\name{lessLM-package}
\alias{lessLM-package}
\alias{lessLM}
\docType{package}
\title{\packageTitle{lessLM}}
\description{\packageDescription{lessLM}}
\section{Package Content}{\packageIndices{lessLM}}
\author{\packageAuthor{lessLM}}
\section{Maintainer}{\packageMaintainer{lessLM}}
\keyword{package}
